--- components/unirate/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@pipedream/unirate",
-  "version": "0.1.0",
-  "description": "Pipedream UniRate Components",
-  "main": "unirate.app.mjs",
-  "keywords": [
-    "pipedream",
-    "unirate"
-  ],
-  "homepage": "https://pipedream.com/apps/unirate",
-  "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@pipedream/platform": "^3.1.1"
-  }
-}
