+++ components/unirateapi/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/unirateapi",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Pipedream UniRateAPI Components",
   "main": "unirateapi.app.mjs",
   "keywords": [
@@ -11,5 +11,8 @@
   "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "@pipedream/platform": "^3.4.0"
   }
 }
